microservices/api_interface/requirements.txt
flask ~= 2.2.3
grpcio-tools ~= 1.30
Jinja2 ~= 3.1.2
grpc-interceptor == 0.11.0
protobuf==5.29.0